feat: implement dynamic reverse proxy configuration system and route parser
- Add complete configuration-driven reverse proxy module
  - Support dynamic loading of proxy targets (IP+port) from config files
  - Add ReverseClientConfig structure for proxy settings
  - Implement hot application of config to client connections

- Refactor route parsing system
  - Implement StateNode::parse_from_str() for path expression parsing
  - Support dynamic parameter capture like '/api/v1/{param}'
  - Add escape character handling ('\{' → '{')

- Fix critical bugs and improvements
  - Fix missing return statement in Lambda expressions
  - Improve URL parser to handle empty paths and relative paths
  - Standardize on size_t instead of unsigned int to prevent overflow
  - Add quoted string handling for config value extraction

- Enhance configuration file syntax
  - Support nested configuration blocks and string escaping
  - Add modular configuration loading mechanism
  - Implement flexible route-to-module mapping

Configuration example:
  location "/api/v1/{path}" {
      type proxy;
      target {
          name 127.0.0.1;
          port 7000;
      };
  }

+++ src/config.cpp
@@ -145,7 +145,7 @@ Config::LoadResult Config::analyse_words(std::vector<std::string> & tokens){
     return LoadResult::OK;
 }
 
-std::optional<std::reference_wrapper<Config::Node>> Config::Node::get_node_recursive(const std::vector<std::string_view> & location,unsigned int expected_index){
+std::optional<std::reference_wrapper<Config::Node>> Config::Node::get_node_recursive(const std::vector<std::string_view> & location,size_t expected_index){
     Node * current = this;
     size_t i = 0,cmp_index = 0;
     size_t founded = 0;
